move prior functions from log_impl.h
